mgmtd: Fix memleak in mgmt frontend client.
During daemon startup if the client and server
is not connected, the session create request
returns false and the session that was created
gets stale in the session list.

So if session create request returns false, free
up the memory.
Along with this, when the client lib is destroyed,
free up any remaining sessions present in the list.

Signed-off-by: Yash Ranjan <ranjany@vmware.com>