mgmtd: simplify config state machine
Simplify the configuration commit process.

- Remove an explicit TXN create and delete handshake with the backend daemons.
Now we just pass a txn-id when we send the config -- the backends can create
there own txn tracker with the given config -- and then we send either "apply"
or "delete" (abort) with that txn-id. The backend will cleanup it's txn state
then.

So instead of 4 exchanges (create-txn, send-cfg, apply-cfg, delete-txn) we have
2 (send-cfg, apply-or-abort-cfg).

- We only expect an error during config send for validation, any other error
from a backend (e.g., can't send a message to it) will simply disconnect that
backend.

There is a deeper issue here with validation overall, what happens when the
daemon isn't running to validate the config that it would normally reject? This
config will then be rejected when we try and load the config when that daemon
initializes. This will have to be solved with some re-architecting not in this
cleanup commit.

- Remove bunches of process exit points in the code. Instead return the error
back while unwinding and then finish the TXN in one place.

- Don't fail configurations b/c of unexpected errors from any single
backend/daemon, instead disconnect that backend -- it will reconnect and
resync.

- Don't fail an in-progress configuration operation when a single backend daemon
disconnects (e.g., maybe it crashed or we disconnected it). Just advance the
config process if it was waiting on this backend or do nothing otherwise.

- Don't use an secondary client array of state values for checking for overall
state transition. Just use a bitmask to track when we've gotten back our
replies, and then transition the overall state.

- Refactor a bit code for better re-use and less multiple same checks.

* A future commit will move some of the code around to like-function areas,
rename/standardize variable names, add some locals to simplify access. All in
the name of maintainability. That commit will not change any actual
functionality though.

Signed-off-by: Christian Hopps <chopps@labn.net>